Revert "Re-organize"
This reverts commit b0fdbafdbd2271669f2e48725f1cff3ff48afe47.

# Conflicts:
#	CaPPMS/Data/DBOperations.cs
